Refactor work
