+++ core/chains/evm/client/simulated_backend_client.go
@@ -18,7 +18,6 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/smartcontractkit/chainlink/core/assets"
-	"github.com/smartcontractkit/chainlink/core/chains/evm/gas"
 	evmtypes "github.com/smartcontractkit/chainlink/core/chains/evm/types"
 	"github.com/smartcontractkit/chainlink/core/logger"
 	"github.com/smartcontractkit/chainlink/core/utils"
@@ -462,8 +461,8 @@ func (c *SimulatedBackendClient) BatchCallContext(ctx context.Context, b []rpc.B
 			b[i].Error = err
 		case "eth_getBlockByNumber":
 			if _, ok := elem.Result.(*evmtypes.Head); !ok {
-				if _, ok = elem.Result.(*gas.Block); !ok {
-					return errors.Errorf("SimulatedBackendClient expected return type of [*evmtypes.Head] or [*gas.Block] for eth_getBlockByNumber, got type %T", elem.Result)
+				if _, ok = elem.Result.(*evmtypes.Block); !ok {
+					return errors.Errorf("SimulatedBackendClient expected return type of [*evmtypes.Head] or [*evmtypes.Block] for eth_getBlockByNumber, got type %T", elem.Result)
 				}
 			}
 			if len(elem.Args) != 2 {
