+++ core/chains/evm/gas/chain_specific.go
@@ -1,10 +1,13 @@
 package gas
 
-import "github.com/smartcontractkit/chainlink/core/config"
+import (
+	evmtypes "github.com/smartcontractkit/chainlink/core/chains/evm/types"
+	"github.com/smartcontractkit/chainlink/core/config"
+)
 
 // chainSpecificIsUsable allows for additional logic specific to a particular
 // Config that determines whether a transaction should be used for gas estimation
-func (tx *Transaction) chainSpecificIsUsable(cfg Config) bool {
+func chainSpecificIsUsable(tx evmtypes.Transaction, cfg Config) bool {
 	if cfg.ChainType() == config.ChainXDai {
 		// GasPrice 0 on most chains is great since it indicates cheap/free transactions.
 		// However, xDai reserves a special type of "bridge" transaction with 0 gas
